Connection sharing - allows a single rdesktop connection to launch multiple applications.
When run in seamless mode, rdesktop creates and listens on a control
socket. A new option allows rdesktop to be run in slave mode, which
notifies the master rdesktop instance of a new command to be run and
then exits. The master instance sends a client-to-server message to the
SeamlessRDP server component, which runs the new command.

http://www.fontis.com.au/rdesktop
https://github.com/pengliang/rdesktop-seamless